script-parse, new script
i record a lot of interactive sessions using `script', that will
save all the output in a log file. the log file however contains
escape sequences for backspace etc., so the log file can be quiet
unreadable.

this script parses the original log file, and returns the readable
version of the interactive session (same output in the terminal
as it would be todo `cat' on the log file).