git clone wrapper
a wrapper around git clone, that will configure the repository
user.name and user.email (since ~/.gitconfig does not contain
any defaults).

if GIT_USER_NAME or GIT_USER_EMAIL is set (preferably in
~/.bashrc.local), the value of these will be used.

if `git clone --use-defaults <repo> [<dir>]` is used, then the
default friendly name for your user account will be used for
user.name, and a combination of $USER and `hostname -f` will
be used for user.email.

if neither is specified, you'll be asked to enter the values you
want for each setting. if you don't enter anything and just hit
enter, then the default "calculated" values will be used
(presented in the prompt question).